Fix encoder bug when flattening nested types.
For a nested type that is flattened during encoding, the outer type and the inner type both create a container at the same nesting level. However, `MessagePackEncoder` does not currently support this use case because it does the encoding when a container is deinitialized.

This change allows containers at the same nesting level to share the same underlying storage and defers the encoding until the latest point in the flow.