Fix decoder bug when unflattening unkeyed message into nested types.
Currently, `MessagePackDecoder` does the decoding when a caller requests a container. Decoding results are not cached.

For a flattened message that is decoded into nested types, the outer type and the inner type both create a container at the same nesting level. For an unkeyed message, the current approach does not work because the unkeyed container returned to the outer type and the inner type need to share the same underlying storage so that the current array index stays consistent.

This change allows containers at the same nesting level to share the same underlying storage, which includes the current array index for unkeyed containers. This approach also has the performance benefit of caching the result of the container split.