Verify the first byte when decoding `Double`/`Float`.
Currently, the decoding code for `Double`/`Float` parses the data without first verifying that the data represents a `Double`/`Float`. This is unsafe. For example, a value of type `UInt64` could be decoded as a `Double` with the same bit pattern (but not the same semantic value).

Instead, the decoding code should always check the first byte before continuing to parse the data.